requirements.txt
torch>=2.0
torchvision>=0.15
wandb
trimesh
pyyaml
matplotlib
numpy
Pillow
scipy
tqdm
